fix: move GraphQLClient to lodgingconnectivity package
